feat(Backend): update

+++ BackEnd/InventoryService/src/main/java/com/example/inventoryservice/Client/WarehouseService/Dto/Form/UpdateOccupancyRequest.java
@@ -4,11 +4,13 @@ import jakarta.persistence.Entity;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Builder
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class UpdateOccupancyRequest{
-    Integer occupancyChange;
+    BigDecimal occupancyChange;
 }
