+++ BackEnd/InventoryService/src/main/java/com/example/inventoryservice/Dtos/Response/InventoryProductResponseNoList.java
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -20,7 +21,7 @@ import java.util.List;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class InventoryProductResponseNoList {
     String inventoryProductId;
-    Integer totalQuantity;
+    BigDecimal totalQuantity;
     Integer minStockLevel;
     Integer maxStockLevel;
     LocalDateTime lastImportDate;
