+++ BackEnd/Order/src/main/java/com/example/order/Client/Inventory/Dto/Resquest/InventoryWarehouseRequest.java
@@ -2,6 +2,7 @@ package com.example.order.Client.Inventory.Dto.Resquest;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Builder
@@ -9,7 +10,7 @@ public record InventoryWarehouseRequest(
         String product,
         String warehouse,
         String bin,
-        Integer quantity,
+        BigDecimal quantity,
         LocalDate expiryDate,
         String status
 ) {}
